Added handling of 'sbmdebugger' VH messages. Example:
sbmdebugger 123 1 request string-array "ret = scene.getCharacterNames()"

will return:


scene.setScale() will now control size of grid lines in viewer.
Added "scale" attribute to SBScene.
Fixed capitalization of included file for Linux builds.

+++ trunk/core/smartbody/smartbody-lib/src/sbm/SBScene.h
@@ -28,8 +28,8 @@ class SBScene : public SBObject
 		SBScene(void);
 		~SBScene(void);
 
-		void setScale(double val);
-		double getScale();
+		void setScale(float val);
+		float getScale();
 
 		void command(const std::string& command);
 		void commandAt(float seconds, const std::string& command);
@@ -108,7 +108,7 @@ class SBScene : public SBObject
 		SBParser* _parser;
 
 		std::map<std::string, SBScript*> _scripts;
-		double _scale;
+		float _scale;
 
 		SbmDebuggerServer* _debuggerServer;
 };
